Remove deprecated dependency: Textrue's Embeddium Options
Embeddium 3.16+ adds similar functionality to TEO, and as such the mod is discontinued and Embeddium no longer supports loading it. As such, Embeddium/Magnesium Extras no longer works as is.

The modification here allows it to still work; all it does is remove the deprecated dependency. The Embeddium dependency should be updated to require version 3.16+ alongside this change.